mybatis-plus目前基本框架搭建完毕

+++ common/src/main/java/com/gree/entity/po/PermissionPO.java
@@ -52,7 +52,7 @@ public class PermissionPO extends BaseEntity {
      * 版本号
      */
     @TableField("VERSION")
-    private Double version;
+    private Integer version;
 
     public PermissionPO(String dsca) {
         this.dsca = dsca;
